code/project_threads_llm/final_version/pyproject.toml
[project]
name = "llm-to-tg-final-version"
version = "1.0.0"
description = "Simple bot with LLM"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "structlog>=25.5.0",
]
